+++ src/App.jsx
@@ -19,6 +19,8 @@ import PurchaseDetailsPage from "./pages/PurchaseDetailsPage";
 import StaffPanelPage from "./pages/StaffPanelPage";
 import TrafficPage from "./pages/TrafficPage";
 import NotFoundPage from "./pages/NotFoundPage";
+import PrivacyPage from "./pages/PrivacyPage";
+import TermsPage from "./pages/TermsPage";
 import ScrollToTop from "./components/common/ScrollToTop.jsx";
 
 export default function App() {
@@ -38,6 +40,8 @@ export default function App() {
           <Route path="purchase/:orderId" element={<PurchaseDetailsPage />} />
           <Route path="staff" element={<StaffPanelPage />} />
           <Route path="traffic" element={<TrafficPage />} />
+          <Route path="privacy" element={<PrivacyPage />} />
+          <Route path="terms" element={<TermsPage />} />
           <Route path="*" element={<NotFoundPage />} />
         </Route>
         <Route path="/login" element={<LoginPage />} />